requirements.txt
django 
djangorestframework
drf-yasg

